refactor: move conditions in separate packages
