+++ pkg/condition/github/github_test.go
@@ -1,4 +1,4 @@
-package condition
+package github
 
 import (
 	"os"
@@ -10,6 +10,6 @@ import (
 func TestGithubValid(t *testing.T) {
 	gha := GitHubActions{}
 	os.Setenv("GITHUB_REF", "")
-	err := gha.RunCondition(CIConfig{"defaultBranch": ""})
+	err := gha.RunCondition(map[string]interface{}{"defaultBranch": ""})
 	assert.EqualError(t, err, "This test run is not running on a branch build.")
 }
