+++ pkg/condition/gitlab/gitlab.go
@@ -1,4 +1,4 @@
-package condition
+package gitlab
 
 import (
 	"fmt"
@@ -24,7 +24,7 @@ func (gl *GitLab) IsBranchRef() bool {
 	return gl.GetCurrentBranch() != ""
 }
 
-func (gl *GitLab) RunCondition(config CIConfig) error {
+func (gl *GitLab) RunCondition(config map[string]interface{}) error {
 	defaultBranch := config["defaultBranch"].(string)
 	if !gl.IsBranchRef() {
 		return fmt.Errorf("This test run is not running on a branch build.")
