+++ pkg/condition/travis/travis_test.go
@@ -1,4 +1,4 @@
-package condition
+package travis
 
 import (
 	"os"
@@ -10,6 +10,6 @@ import (
 func TestTravisValid(t *testing.T) {
 	travis := TravisCI{}
 	os.Setenv("TRAVIS", "")
-	err := travis.RunCondition(CIConfig{"token": "", "defaultBranch": "", "private": false})
+	err := travis.RunCondition(map[string]interface{}{"token": "", "defaultBranch": "", "private": false})
 	assert.EqualError(t, err, "semantic-release didn’t run on Travis CI and therefore a new version won’t be published.")
 }
